refactor(serviceevents): make the endpoint span processor the sole endpoint path
Remove OTEL_AWS_SERVICE_EVENTS_USE_SPAN_PROCESSOR and the legacy per-framework
hooks. The framework-agnostic ServiceEventsSpanProcessor (already the default)
is now the only endpoint-measurement path, matching the Java and Python distros.

- Delete the per-framework hook modules (Express/Fastify/Koa/Next.js) plus the
  global HTTP patch and unmatched-route helper, and their tests.
- Relocate extractErrorFromCallPath into a dedicated processor/error-extraction
  module (it only reads monitor investigation data — no framework dependency).
- Drop the now-orphaned config knobs: useSpanProcessor and the internal
  instrumentExpress/Fastify/Koa/NextJs toggles.
- initialize() installs the span processor unconditionally; if registration on
  the active tracer provider fails it warns and emits no endpoint signals
  (there is no longer a hook fallback to degrade to).

Endpoint signals are byte-identical — only the internal mechanism changed.
