+++ aws-distro-opentelemetry-node-autoinstrumentation/src/serviceevents/index.ts
@@ -20,4 +20,3 @@ export * from './models';
 export * from './utils';
 export * from './collectors';
 export * from './exporter';
-export * from './instrumentation';
